+++ M2/Macaulay2/e/imonorder.hpp
@@ -6,19 +6,7 @@
 /* This is the internal form of the monomial ordering */
 /* Used in monomial encoding/decoding/comparison */
 
-#include "engine-includes.hpp"
-
-#ifdef HAVE_ALLOCA_H
-#include <alloca.h>
-#else
-#ifdef HAVE_MALLOC_H
-// under minggw32, alloca is declared in malloc.h
-#include <malloc.h>
-#endif
-#endif
-
-#include "interface/random.h"
-#include "monordering.h"
+#include "interface/monomial-ordering.h"
 
 typedef int *exponents;
 typedef int *monomial;
